+++ hutool-core/src/main/java/cn/hutool/core/collection/CollUtil.java
@@ -2128,7 +2128,7 @@ public class CollUtil {
 		int resultSize = list.size();
 		// 每页条目数大于总数直接返回所有
 		if (resultSize <= pageSize) {
-			if (pageNo <= 1) {
+			if (pageNo < 1) {
 				return Collections.unmodifiableList(list);
 			} else {
 				// 越界直接返回空