+++ hutool-db/src/main/java/cn/hutool/db/Page.java
@@ -1,11 +1,11 @@
 package cn.hutool.db;
 
-import java.io.Serializable;
-import java.util.Arrays;
-
 import cn.hutool.core.util.ArrayUtil;
 import cn.hutool.core.util.PageUtil;
 import cn.hutool.db.sql.Order;
+
+import java.io.Serializable;
+import java.util.Arrays;
 
 /**
  * 分页对象
@@ -50,11 +50,11 @@ public class Page implements Serializable {
 	 * 构造
 	 * 
 	 * @param pageNumber 页码，0表示第一页
-	 * @param numPerPage 每页结果数
+	 * @param pageSize 每页结果数
 	 * @param order 排序对象
 	 */
-	public Page(int pageNumber, int numPerPage, Order order) {
-		this(pageNumber, numPerPage);
+	public Page(int pageNumber, int pageSize, Order order) {
+		this(pageNumber, pageSize);
 		this.orders = new Order[] { order };
 	}
 	// ---------------------------------------------------------- Constructor start
@@ -109,7 +109,7 @@ public class Page implements Serializable {
 	 * @param pageSize 每页结果数
 	 */
 	public void setPageSize(int pageSize) {
-		this.pageSize = pageSize <= 0 ? DEFAULT_PAGE_SIZE : pageSize;
+		this.pageSize = (pageSize <= 0) ? DEFAULT_PAGE_SIZE : pageSize;
 	}
 
 	/**
@@ -145,14 +145,14 @@ public class Page implements Serializable {
 	 * @return 开始位置
 	 */
 	public int getStartPosition() {
-		return getStartEnd()[0];
+		return PageUtil.getStart(this.pageNumber, this.pageSize);
 	}
 
 	/**
 	 * @return 结束位置
 	 */
 	public int getEndPosition() {
-		return getStartEnd()[1];
+		return PageUtil.getEnd(this.pageNumber, this.pageSize);
 	}
 
 	/**