+++ hutool-setting/src/test/java/cn/hutool/setting/test/SettingUtilTest.java
@@ -14,6 +14,12 @@ public class SettingUtilTest {
 	}
 
 	@Test
+	public void getTest2() {
+		String driver = SettingUtil.get("example/example").get("demo", "key");
+		Assert.assertEquals("value", driver);
+	}
+
+	@Test
 	public void getFirstFoundTest() {
 		String driver = SettingUtil.getFirstFound("test2", "test").get("demo", "driver");
 		Assert.assertEquals("com.mysql.jdbc.Driver", driver);